requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
streamlit
pillow